Fixed code and test to make it pass
Summary: I've added the "onclick" ability so when user clicks "continue" it will go to the next page. In the same time I've fixed the tests so they include ".click()" and also the correct address for the next page.

Tasks: I've discovered another issue that is cause by how you styled your button to align to right. The best way to understand is that you just test it yourself - click on the "continue" button. This issue is caused on line 99 with following code: style="position: absolute; right: 30%;"