Create registration list (#17)
* add registration list page

* add button, service name and more tests

* add button and test

* Fixed code and test to make it pass

Summary: I've added the "onclick" ability so when user clicks "continue" it will go to the next page. In the same time I've fixed the tests so they include ".click()" and also the correct address for the next page.

Tasks: I've discovered another issue that is cause by how you styled your button to align to right. The best way to understand is that you just test it yourself - click on the "continue" button. This issue is caused on line 99 with following code: style="position: absolute; right: 30%;"

* Button fix

Summary: Styling used for button to be align to right caused it to jump to header when clicked on and the click function wouldn't work. That is now fixed.

Tasks:

* add DS_Store files to .gitignore

* added new lines to the end of files

* Delete .DS_Store

* Delete .DS_Store

* remove index.html from links for Continue button

* add link to back button

* add registration-list directory

* Remove .html from links in tests

Summary:

Tasks:

Co-authored-by: Miroslav Palka <palkamiroslav.v@gmail.com>
Co-authored-by: Miroslav Palka <Porim@users.noreply.github.com>