Create 'Upload choice' page
Summary: This is the page where users specify which document they want to
upload. It consists of a title, subtitle and a couple of radios as per
the design and has tests to check if the content is correct.

Tasks: Update the links for the back and continue buttons, so they link
to the right pages.